Revert "Revert nextjs 14 upgrade (#3036)"
This reverts commit f45f40f9eb42945253c238d0ebbdcb8bf04bd36a.
