+++ .env.test.local.example
@@ -10,3 +10,4 @@ SERVERLESS_PORT=3020
 MAILGUN_DOMAIN=mg.charmverse.com
 MAILGUN_KEY=test-key
 PERMISSIONS_API_URL=http://127.0.0.1:3001
+REACT_APP_WEBSOCKETS_HOST=http://127.0.0.1:3336
