+++ adapters/redis/redisClient.ts
@@ -2,17 +2,12 @@ import { log } from '@charmverse/core/log';
 import type { RedisClientType } from 'redis';
 import { createClient } from 'redis';
 
-// Export the singleton instance
-declare global {
-  // eslint-disable-next-line no-var, vars-on-top
-  var redisClient: RedisClientType;
-}
-
 let redisClientInstance: RedisClientType | null = null;
 
 if (process.env.REDIS_URI) {
   try {
-    redisClientInstance = global.redisClient ?? createClient({ url: process.env.REDIS_URI });
+    redisClientInstance =
+      ((global as any).redisClient as RedisClientType) ?? createClient({ url: process.env.REDIS_URI });
     redisClientInstance.on('error', (error) => log.error('Redis Client Error', error));
   } catch (error) {
     log.error('Could not instantiate Redis. Error occurred', error);
@@ -21,7 +16,7 @@ if (process.env.REDIS_URI) {
 
 // remember this instance of prisma in development to avoid too many clients
 if (process.env.NODE_ENV === 'development' && redisClientInstance) {
-  global.redisClient = redisClientInstance;
+  (global as any).redisClient = redisClientInstance;
 }
 
 export const redisClient = redisClientInstance;
