+++ package.json
@@ -15,13 +15,13 @@
     "cron:staging": "npx dotenv -e .env.staging -- node ./.next/server/cron.js",
     "cron:prod": "node ./.next/server/cron.js",
     "cron:dev": "npx dotenv -e .env.local -- ts-node --transpileOnly ./background/cron.ts",
-    "sockets": "concurrently --names \"next,sockets\" -c \"green,yellow\" \"npm run server:sockets\" \"npm run sockets:dev\"",
     "sockets:build": "tsc --project tsconfig.websockets.json && tsc-alias -p tsconfig.websockets.json",
     "sockets:esbuild": "esbuild background/initWebsockets.ts --bundle --outdir=dist --tsconfig=tsconfig.websockets.json",
-    "sockets:dev": "cross-env PORT=3002 npx dotenv -e .env.local -- node --watch ./.next/server/websockets.js",
-    "sockets:dev:esbuild": "cross-env PORT=3002 npx dotenv -e .env.local -- tsx watch --tsconfig tsconfig.websockets.json ./background/initWebsockets.ts",
-    "sockets:prod": "node ./.next/server/websockets.js",
+    "sockets:dev": "cross-env PORT=3336 npx dotenv -e .env.local -- node --watch ./.next/server/websockets.js",
+    "sockets:dev:esbuild": "cross-env PORT=3336 npx dotenv -e .env.local -- tsx watch --tsconfig tsconfig.websockets.json ./background/initWebsockets.ts",
     "sockets:prod:esbuild": "node dist/background/initWebsockets.js",
+    "sockets:test": "dotenv -e .env.test.local -v PORT=3336 -- node ./.next/server/websockets.js",
+    "sockets:test:ci": "npm run sockets:test",
     "sockets:staging": "npx dotenv -e .env.staging -- node ./.next/server/websockets.js",
     "prisma:migrate": "cd node_modules/@charmverse/core && npm run prisma:migrate",
     "prisma:studio": "cd node_modules/@charmverse/core && npm run prisma:studio",
@@ -33,15 +33,16 @@
     "postinstall": "patch-package && cd node_modules/@charmverse/core && npm run postinstall",
     "migrate:test": "dotenv -e .env.test.local -- npx prisma migrate deploy",
     "notifications:debug": "ts-node --transpileOnly ./scripts/readNotifications.ts",
-    "server:dev": "npx react-env --path .env.local -- node node_modules/.bin/next start",
-    "server:sockets": "cross-env REACT_APP_WEBSOCKETS_HOST=http://localhost:3002 npm start",
-    "start": "cross-env NODE_OPTIONS=\"--max_old_space_size=8192 --experimental-vm-modules\" npx react-env --path .env.local -- next dev",
+    "server:dev": "NODE_OPTIONS=\"--max_old_space_size=8192\" npx react-env --path .env.local -- next dev",
+    "server:test": "npx react-env --path .env.test.local -- next dev",
+    "server:test:ci": "NODE_OPTIONS=\"--max_old_space_size=4096\" npx react-env --path .env.test.local -- npx next start",
+    "start": "concurrently --names \"next,sockets\" -c \"green,yellow\" \"npm run server:dev\" \"npm run sockets:dev\"",
     "start:debug": "cross-env NODE_OPTIONS='--inspect' next",
     "start:ssl": "(npx local-ssl-proxy --source 3000 --target 8080) & next dev --port 8080",
     "start:ssl:prod": "(npx local-ssl-proxy --source 3000 --target 8080) & next start --port 8080",
-    "start:test": "cross-env NODE_OPTIONS=\"--max_old_space_size=4096 --experimental-vm-modules\"  react-env --path .env.test.local -- npm start",
-    "start:test:ci": "NODE_OPTIONS=\"--max_old_space_size=4096\" npx react-env --path .env.test.local -- npx next start",
-    "start:remote": "cross-env NODE_ENV=development npx react-env --path .env.production -- npm start",
+    "start:test": "concurrently --names \"next,sockets\" -c \"green,yellow\" \"npm run server:test\" \"npm run sockets:test\"",
+    "start:test:ci": "concurrently --names \"next,sockets\" -c \"green,yellow\" \"npm run server:test:ci\" \"npm run sockets:test:ci\"",
+    "start:remote": "concurrently --names \"next,sockets\" -c \"green,yellow\" \"npx react-env --path .env.production -- npm run server:dev\" \"npm run sockets:dev\"",
     "start:staging": "npm run prisma:deploy && next start",
     "prepare": "husky install",
     "db-tool": "npx prisma studio",
@@ -122,7 +123,7 @@
     "@mux/mux-node": "^7.0.0",
     "@mux/mux-video-react": "^0.7.15",
     "@mux/upchunk": "^3.0.0",
-    "@next/bundle-analyzer": "^13.5.2",
+    "@next/bundle-analyzer": "^13.5.6",
     "@notionhq/client": "^2.2.2",
     "@react-dnd/shallowequal": "^4.0.0",
     "@react-email/column": "0.0.7",
@@ -143,6 +144,7 @@
     "@safe-global/safe-ethers-lib": "^1.9.4",
     "@safe-global/safe-service-client": "^2.0.2",
     "@snapshot-labs/snapshot.js": "^0.4.60",
+    "@socket.io/postgres-adapter": "^0.3.1",
     "@socket.io/redis-adapter": "^8.2.1",
     "@storybook/icons": "^1.2.1",
     "@stripe/react-stripe-js": "^2.1.1",
@@ -187,11 +189,12 @@
     "msw-storybook-addon": "^1.10.0",
     "nanoid": "^3.3.6",
     "nanoid-dictionary": "^3.0.0",
-    "next": "^13.2.2",
+    "next": "^14.0.1",
     "next-connect": "^0.13.0",
     "next-swagger-doc": "^0.3.6",
     "node-cron": "^3.0.0",
     "papaparse": "^5.3.2",
+    "pg": "^8.11.3",
     "promise-retry": "^2.0.1",
     "prosemirror-changeset": "^2.2.0",
     "prosemirror-collab": "^1.3.0",
@@ -252,14 +255,14 @@
   "devDependencies": {
     "@depay/web3-mock": "^13.18.1",
     "@lit-protocol/types": "^2.2.61",
-    "@next/eslint-plugin-next": "^12.0.10",
+    "@next/eslint-plugin-next": "^13.5.6",
     "@playwright/test": "^1.39.0",
     "@storybook/addon-essentials": "^7.5.1",
     "@storybook/addon-interactions": "^7.5.1",
     "@storybook/addon-links": "^7.5.1",
     "@storybook/blocks": "^7.5.1",
     "@storybook/manager-api": "^7.5.1",
-    "@storybook/nextjs": "^7.5.1",
+    "@storybook/nextjs": "^7.5.2",
     "@storybook/react": "^7.5.1",
     "@storybook/testing-library": "^0.2.2",
     "@storybook/theming": "^7.5.1",
@@ -345,7 +348,7 @@
     "storybook": "^7.5.1",
     "supertest": "^6.3.3",
     "svgdom": "^0.1.16",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tsc-alias": "^1.8.6",
     "tsconfig-paths": "^4.2.0",
